feat: modernize usage of setuptools
- use pyproject.toml consistently
- move c_ext.py into setup.py to avoid import error
- remove deprecated usage of distutils
- unconditional install_requires

also:
- fix build cache usage
- script style changes
- remove unused coverage dependency
