+++ .github/workflows/run-tests.yaml
@@ -61,7 +61,7 @@ jobs:
         id: restore_gk_pkg
         uses: actions/cache/restore@v4
         with:
-          key: mamba-env-${{ runner.os }}-${{ runner.arch }}-${{ hashFiles('.github/workflows/run-tests.yaml', '.github/actions/**', 'conda-recipe/**', 'src/**', 'genome_kit/**', 'setup.py', 'setup/**', 'tests/**') }}
+          key: gk-tarballs-${{ runner.os }}-${{ runner.arch }}-${{ hashFiles('.github/workflows/run-tests.yaml', '.github/actions/**', 'conda-recipe/**', 'src/**', 'genome_kit/**', 'setup.py', 'setup/**', 'tests/**') }}
           path: |
             ~/conda-bld
           fail-on-cache-miss: true
@@ -74,10 +74,12 @@ jobs:
       - name: create test env
         shell: bash -l -e {0}
         run: |
+          # boa required for the mambabuild subcommand
           set -x
+          # inconsistently got "coverage 7.9.1 is not supported on this platform" so removed
+          # (not needed?)
           if [ ! -d "${HOME}/micromamba/envs/test" ]; then
-            micromamba create -yqn test -c conda-forge \
-              boa==0.16.0 mamba==1.5.7 conda==24.1.2 coverage
+            micromamba create -yqn test -c conda-forge boa==0.16.0 mamba==1.5.7 conda==24.1.2
           fi
           set +x
