+++ pyproject.toml
@@ -1,0 +1,7 @@
+[build-system]
+requires = [
+    "setuptools>=61",
+    "wheel",
+    "numpy==2"
+]
+build-backend = "setuptools.build_meta"
